feat: role fixed number to string
