+++ lib/next-auth/utils.ts
@@ -1,5 +1,5 @@
 import { ROLE_ADMIN } from '@/constants';
 
-export function checkAdmin(role: number | null): boolean {
+export function checkAdmin(role: string | null): boolean {
   return !!role && role === ROLE_ADMIN;
 }
