+++ types/next-auth.d.ts
@@ -2,13 +2,13 @@ import { DefaultSession, Session, User } from 'next-auth';
 import { JWT } from 'next-auth/jwt';
 
 type UserId = string;
-type RoleId = number;
+type Role = string;
 
 declare module 'next-auth' {
   interface Session {
     user: {
       id: UserId | null | undefined;
-      role: RoleId | null | undefined;
+      role: Role | null | undefined;
     } & DefaultSession['user'];
   }
 }
@@ -16,7 +16,7 @@ declare module 'next-auth' {
 declare module 'next-auth/jwt' {
   interface JWT {
     id?: UserId;
-    role?: RoleId;
+    role?: Role;
     exp?: number;
   }
 }
